+++ lib/additionals/wiki_macros/project_macro.rb
@@ -36,11 +36,7 @@ module Additionals
   end
 
   def self.load_projects
-    all_projects = if ActiveRecord::VERSION::MAJOR < 4
-                     Project.active.visible(User.current).find(:all, order: 'projects.name')
-                   else
-                     Project.active.visible.sorted
-                   end
+    all_projects = Project.active.visible.sorted
     my_projects = []
     all_projects.each do |p|
       my_projects << p if User.current.member_of?(p)