+++ test/test_helper.rb
@@ -38,11 +38,7 @@ module Additionals
     end
 
     def self.create_fixtures(fixtures_directory, table_names, _class_names = {})
-      if ActiveRecord::VERSION::MAJOR >= 4
-        ActiveRecord::FixtureSet.create_fixtures(fixtures_directory, table_names, _class_names = {})
-      else
-        ActiveRecord::Fixtures.create_fixtures(fixtures_directory, table_names, _class_names = {})
-      end
+      ActiveRecord::FixtureSet.create_fixtures(fixtures_directory, table_names, _class_names = {})
     end
 
     def self.prepare